+++ railties/test/application/configuration_test.rb
@@ -2184,6 +2184,19 @@ module ApplicationTests
       assert_ractor_shareable ActionDispatch::Request.parameter_parsers
     end
 
+    test "action_view configuration is made Ractor-shareable when unshareable_proc_action is set" do
+      remove_from_config '.*config\.load_defaults.*\n'
+
+      app_file "config/initializers/freeze_configuration.rb", <<-RUBY
+        ActiveSupport::Ractors.unshareable_proc_action = :raise
+      RUBY
+
+      app "production"
+
+      assert_ractor_shareable ActionView::Base.field_error_proc
+      assert_ractor_shareable ActionView::Base.default_formats
+    end
+
     test "config.annotations wrapping SourceAnnotationExtractor::Annotation class" do
       make_basic_app do |application|
         application.config.annotations.register_extensions("coffee") do |tag|